style(log): 일별 및 시간별 운행기록 모바일 반응형 작업 (#117)
