+++ app/(main)/log/[id]/daily/styles.css.ts
@@ -1,6 +1,6 @@
 import { style } from '@vanilla-extract/css'
 
-import { vars } from '@/styles/theme.css'
+import { breakPoints, vars } from '@/styles/theme.css'
 
 export const container = style({
     width: '100%',
@@ -9,6 +9,20 @@ export const container = style({
     gap: '34px',
     padding: '30px',
     overflowY: 'auto',
+
+    '@media': {
+        [`screen and (max-width: ${breakPoints.mobile}px)`]: {
+            marginBottom: '70px',
+        },
+    },
+})
+
+export const breadcrumbsWrapper = style({
+    '@media': {
+        [`screen and (max-width: ${breakPoints.mobile}px)`]: {
+            display: 'none',
+        },
+    },
 })
 
 export const contents = style({
@@ -17,6 +31,12 @@ export const contents = style({
     alignItems: 'flex-start',
     gap: '30px',
     marginTop: '30px',
+
+    '@media': {
+        [`screen and (max-width: ${breakPoints.mobile}px)`]: {
+            flexDirection: 'column',
+        },
+    },
 })
 
 export const contentWrapper = style({
@@ -27,8 +47,20 @@ export const contentWrapper = style({
 
 export const leftSection = style({
     width: '50%',
+
+    '@media': {
+        [`screen and (max-width: ${breakPoints.mobile}px)`]: {
+            width: '100%',
+        },
+    },
 })
 
 export const rightSection = style({
     width: '50%',
+
+    '@media': {
+        [`screen and (max-width: ${breakPoints.mobile}px)`]: {
+            width: '100%',
+        },
+    },
 })
